refactor(diet): extract import-path heuristics to import_paths.go
Move stateless import-path helpers and ecosystem data tables out of
service.go into a new same-package file import_paths.go:
  - isWorkspaceDep, filterWorkspaceDeps
  - buildImportPaths
  - pypiPrefixes, buildPyPIImportPaths, isPythonIdentifierSafe,
    isPythonDottedIdentifierSafe
  - mavenPackageOverrides, mavenRuntimeDeps, buildMavenImportPaths,
    isJavaPackageSafe, isJavaDottedPackageSafe

service.go retains only parsePURLParts, the Service methods (Run,
buildEntries, resolveUnmatchedPyPI), and computeHealthSignals.

No caller changes — same-package visibility is preserved.
